test_vision_direct.py
"""
Simple direct test with Gemini Vision API
"""
import os
from dotenv import load_dotenv
import google.generativeai as genai
from PIL import Image, ImageDraw
import io
import base64

load_dotenv()

# Configure Gemini
genai.configure(api_key=os.getenv('GEMINI_API_KEY'))

def create_sky_image():
    """Create a sky-like image"""
    width, height = 800, 600
    img = Image.new('RGB', (width, height))
    pixels = img.load()
    
    # Blue sky gradient
    for y in range(height):
        blue_value = int(135 + (y / height) * 100)
        for x in range(width):
            pixels[x, y] = (100, 150, blue_value)
    
    # Add clouds
    draw = ImageDraw.Draw(img)
    draw.ellipse([100, 100, 250, 180], fill=(255, 255, 255))
    draw.ellipse([200, 120, 320, 190], fill=(255, 255, 255))
    draw.ellipse([500, 150, 650, 220], fill=(255, 255, 255))
    
    return img

print("=" * 70)
print("🖼️  Direct Gemini Vision Test")
print("=" * 70)

# Create test image
print("\n📸 Creating sky image with clouds...")
img = create_sky_image()
print("✅ Image created\n")

# Test with Gemini
print("🔍 Asking Gemini: 'Does this image show a real blue sky with clouds?'\n")
print("-" * 70)

model = genai.GenerativeModel('gemini-2.5-flash')
response = model.generate_content([
    "Analyze this image. Does it show a real blue sky with clouds photographed outdoors? Or is it computer-generated/artificial? Be honest and detailed.",
    img
])

print("🤖 Gemini's Analysis:")
print("-" * 70)
print(response.text)
print("-" * 70)

print("\n✅ Vision API working! Gemini can analyze images.")
print("🎯 The AI correctly identifies artificial vs real images.\n")
print("=" * 70)
