requirements.txt
# Core Flask
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0

# Database
pymongo==4.6.1
mongoengine==0.27.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Google Gemini AI
google-generativeai>=0.8.0

# External APIs
requests==2.31.0
twilio==8.11.0

# Firebase
firebase-admin==6.3.0

# Image processing & AI
Pillow==11.1.0
opencv-python==4.10.0.84
tensorflow==2.20.0
numpy>=2.0.0

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-flask==1.3.0
